project/requirements/base.txt
django-authtools==1.6.0
django-braces==1.13.0
django-crispy-forms==1.7.2
django-environ==0.4.5
django==2.2.2
easy-thumbnails==2.5
pillow==5.3.0
pytz==2018.7
gunicorn==19.9.0
django-adminlte2
django_tables2
django-role-permissions==2.2.1
django-extensions==2.1.5
djangorestframework==3.9.1
uwsgi==2.0.18
psycopg2-binary==2.7.5

